requirements.txt
# Polymarket Trading Bot Dependencies

# Core
requests>=2.28.0
python-dotenv>=1.0.0

# Trading API
py-clob-client>=0.34.0
web3>=7.0.0

# Real-time
websockets>=11.0

# Data Science / Signal Aggregation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# GraphQL (for Goldsky subgraph)
gql[aiohttp]>=3.0.0
